+++ lib/AppInfo/Application.php
@@ -9,7 +9,9 @@ declare( strict_types=1 );
 
 namespace OCA\FileChecksumSearch\AppInfo;
 
+use OCA\FileChecksumSearch\BackgroundJob\DrainPendingUpdates;
 use OCA\FileChecksumSearch\Config\ConfigLexicon;
+use OCA\FileChecksumSearch\Listener\FileListener;
 use OCA\FileChecksumSearch\Migration\LifecycleHandler;
 use OCA\FileChecksumSearch\Search\HashSearchProvider;
 use OCA\FileChecksumSearch\Service\CronJobService;
@@ -20,7 +22,12 @@ use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\AppFramework\Http\Events\BeforeTemplateRenderedEvent;
+use OCP\BackgroundJob\IJobList;
 use OCP\EventDispatcher\IEventDispatcher;
+use OCP\Files\Events\Node\NodeCopiedEvent;
+use OCP\Files\Events\Node\NodeCreatedEvent;
+use OCP\Files\Events\Node\NodeDeletedEvent;
+use OCP\Files\Events\Node\NodeWrittenEvent;
 use OCP\Server;
 use OCP\Util;
 
@@ -46,6 +53,23 @@ class Application
 
 		$context->registerConfigLexicon( ConfigLexicon::class );
 		$context->registerSearchProvider( HashSearchProvider::class );
+
+		$context->registerEventListener(
+			NodeCopiedEvent::class,
+			FileListener::class,
+		);
+		$context->registerEventListener(
+			NodeWrittenEvent::class,
+			FileListener::class,
+		);
+		$context->registerEventListener(
+			NodeCreatedEvent::class,
+			FileListener::class,
+		);
+		$context->registerEventListener(
+			NodeDeletedEvent::class,
+			FileListener::class,
+		);
 	}
 
 
@@ -98,6 +122,11 @@ class Application
 		// Re-register cron job definitions from backup
 		Server::get( CronJobService::class )
 		      ->restore()
+		;
+
+		// Register the pending hash update drain job
+		Server::get( IJobList::class )
+		      ->add( DrainPendingUpdates::class )
 		;
 	}